+++ commands/apache/ApacheRestartCommand.cs
@@ -13,9 +13,7 @@ namespace Gitpod.Tool.Commands.Apache
 
         public override int Execute(CommandContext context, Settings settings)
         {
-            string result = ExecCommand.Exec("apachectl restart");
-
-            AnsiConsole.WriteLine(result);
+            AnsiConsole.WriteLine(ExecCommand.Exec("apachectl restart"));
 
             return 0;
         }
