+++ commands/services/SelectServicesCommand.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Gitpod.Tool.Helper;
-using Gitpod.Tool.Helper.Internal.Config;
+using Gitpod.Tool.Helper.Docker;
+using Gitpod.Tool.Helper.Internal.Config.Sections;
 using Spectre.Console;
 using Spectre.Console.Cli;
