+++ helper/internal/config/sections/ShellScriptConfig.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 
-namespace Gitpod.Tool.Helper.Internal.Config
+namespace Gitpod.Tool.Helper.Internal.Config.Sections
 {
-    class ShellScriptConfig: AbstractConfig
+    class ShellScriptConfig: ConfigHelper
     {
         public static List<string> AdditionalDirectories
         {
