+++ helper/php/PhpHelper.cs
@@ -1,43 +1,20 @@
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Text.RegularExpressions;
-using Gitpod.Tool.Helper.Internal.Config;
+using Gitpod.Tool.Helper.Internal.Config.Sections;
 using Spectre.Console;
 
-namespace Gitpod.Tool.Helper
+namespace Gitpod.Tool.Helper.Php
 {
     class PhpHelper
     {  
-        public static List<string> GetAvailablePhpVersions()
-        {
-            var availablePhpVersions = new List<string>();
-
-            string pattern = @"Alternative: \/usr\/bin\/php([0-9.]+)";
-            string input = ExecCommand.Exec("update-alternatives --query php");
-
-            RegexOptions options = RegexOptions.Multiline;
-        
-            foreach (Match m in Regex.Matches(input, pattern, options))
-            {
-                if (!m.Groups.ContainsKey("1") || m.Groups[1].ToString().Length < 3) {
-                    continue;
-                }
-
-                availablePhpVersions.Insert(0, m.Groups[1].ToString());
-            }
-
-            return availablePhpVersions;
-        }
-
         public static void SetNewPhpVersion(string newVersion, bool isDebug)
         {
             AnsiConsole.Status()
                 .AutoRefresh(true)
                 .Start("Setting PHP Version to " + newVersion, ctx => 
                 {
-                    var availablePhpVersions = GetAvailablePhpVersions();
+                    var availablePhpVersions = PhpVersionHelper.GetAvailablePhpVersions();
 
                     if (!availablePhpVersions.Contains(newVersion)) {
                         AnsiConsole.MarkupLine("Checking if the input is a valid php version....[red]Invalid[/]");
@@ -47,7 +24,7 @@ namespace Gitpod.Tool.Helper
 
                     AnsiConsole.MarkupLine("Checking if the input is a valid php version....[green1]Valid[/]");
 
-                    string currentPhpVersion = GetCurrentPhpVersion();
+                    string currentPhpVersion = PhpVersionHelper.GetCurrentPhpVersion();
 
                     // Check if we have selected another version as the currently active one
                     if (newVersion != currentPhpVersion) {
@@ -78,7 +55,7 @@ namespace Gitpod.Tool.Helper
                         AnsiConsole.MarkupLine("Restarting apache...[green1]Success[/]");
                     }
 
-                    string testResult = PhpHelper.GetCurrentPhpVersionOutput();
+                    string testResult = PhpVersionHelper.GetCurrentPhpVersionOutput();
 
                     if (isDebug) {
                         AnsiConsole.WriteLine(testResult);
@@ -196,192 +173,6 @@ namespace Gitpod.Tool.Helper
                 });
 
             AnsiConsole.MarkupLine("PHP Version has been set to " + newVersion);
-        }
-
-        public static string GetCurrentPhpVersionOutput()
-        {
-            return ExecCommand.Exec("php -version");
-        }
-
-        public static string GetPhpIniPath()
-        {
-            return ExecCommand.Exec("php -i | grep 'Configuration File'");
-        }
-
-        public static string GetCurrentPhpVersion()
-        {
-            string output = GetCurrentPhpVersionOutput();
-
-            Regex regex = new Regex(@"(?:PHP) ([(0-9)].[(0-9)])");
-            Match match = regex.Match(output);
-
-            if (!match.Success) {
-                throw new Exception("Failed to parse the php version command output to find the active version.");
-            }
-
-            return match.Groups[1].Value;
-        }
-
-        public static void UpdatePhpIniFiles(bool isDebug)
-        {
-            string currentPhpVersion = GetCurrentPhpVersion();
-
-            if (isDebug) {
-                AnsiConsole.WriteLine("Active PHP Version " + currentPhpVersion);
-            }
-
-            AnsiConsole.Write("Generating custom.ini....");
-
-            var customIniFiles = GenerateCustomIniFilesFromConfig();
-
-            if (customIniFiles.Count == 0) {
-                AnsiConsole.MarkupLine("[cyan3]No custom settings found[/]");
-
-                return;
-            }
-
-            AnsiConsole.MarkupLine("[green1]Done[/]");
-           
-            if (customIniFiles.ContainsKey("cli")) {
-                AnsiConsole.Write("Copy the custom_cli.ini to the target directory....");
-
-                string targetFile = "/etc/php/" + currentPhpVersion + "/cli/conf.d/custom.ini";
-                
-                if (isDebug) {
-                    AnsiConsole.WriteLine("Copying from \"" + customIniFiles["cli"] + "\" to \"" + targetFile + "\"");
-                }
-                
-                ExecCommand.Exec("sudo cp " + customIniFiles["cli"] + " " + targetFile);
-            }
-
-            if (customIniFiles.ContainsKey("web")) {
-                string targetFile = "/etc/php/" + currentPhpVersion + "/apache2/conf.d/custom.ini";
-                
-                if (isDebug) {
-                    AnsiConsole.WriteLine("Copying from \"" + customIniFiles["web"] + "\" to \"" + targetFile + "\"");
-                }
-                
-                ExecCommand.Exec("sudo cp " + customIniFiles["web"] + " " + targetFile);
-
-                targetFile = "/etc/php/" + currentPhpVersion + "/apache2/conf.d/custom.ini";
-                
-                if (isDebug) {
-                    AnsiConsole.WriteLine("Copying from \"" + customIniFiles["web"] + "\" to \"" + targetFile + "\"");
-                }
-                
-                ExecCommand.Exec("sudo cp " + customIniFiles["web"] + " " + targetFile);
-            }
-
-            AnsiConsole.MarkupLine("[green1]Done[/]");
-        }
-
-        public static void AddSettingToPhpIni(string name, string value, bool setForWeb = false, bool setForCLI = false, bool isDebug = false)
-        {
-            if (name == null || name.Length <= 3 || value == null || value.Length < 1) {
-                AnsiConsole.MarkupLine("[red]Invalid name and/or value[/]");
-
-                return;
-            }
-
-            if (!setForWeb && !setForCLI) {
-                if (PhpConfig.Config.ContainsKey(name)) {
-                    PhpConfig.Config[name] = value;
-                } else {
-                    PhpConfig.Config.Add(name, value);
-                }
-            }
-
-            if (setForWeb) {
-                if (PhpConfig.ConfigWeb.ContainsKey(name)) {
-                    PhpConfig.ConfigWeb[name] = value;
-                } else {
-                    PhpConfig.ConfigWeb.Add(name, value);
-                }
-            }
-
-            if (setForCLI) {
-                if (PhpConfig.ConfigCli.ContainsKey(name)) {
-                    PhpConfig.ConfigCli[name] = value;
-                } else {
-                    PhpConfig.ConfigCli.Add(name, value);
-                }
-            }
-
-            // Update the ini files that are being used by apache and cli
-            PhpHelper.UpdatePhpIniFiles(isDebug);
-
-            if (setForWeb) {
-                ExecCommand.Exec("apachectl restart");
-            }
-        }
-
-        private static Dictionary<string, string> GenerateCustomIniFilesFromConfig()
-        {
-            var applicationDir = AppDomain.CurrentDomain.BaseDirectory;
-            var iniSettingsFolder = applicationDir + "/php";
-            var customFilesWithPath = new Dictionary<string, string>();
-
-            if (!Directory.Exists(iniSettingsFolder)) {
-                Directory.CreateDirectory(iniSettingsFolder);
-            }
-
-            if (PhpConfig.Config.Count == 0 && PhpConfig.ConfigWeb.Count == 0 && PhpConfig.ConfigCli.Count == 0) {
-                return customFilesWithPath;
-            }
-
-            // Combine config and configWeb to a custom.ini for web
-            var configWeb = new Dictionary<string, string>(PhpConfig.ConfigWeb);
-            var customWeb = new Dictionary<string, string>(PhpConfig.Config);
-
-            foreach (KeyValuePair<string, string> item in customWeb) {
-                if (configWeb.ContainsKey(item.Key)) {
-                    customWeb[item.Key] = configWeb[item.Key];
-                    configWeb.Remove(item.Key);
-                }
-            }
-
-            if (configWeb.Count > 0) {
-                customWeb = customWeb.Union(configWeb)
-                                .ToLookup(x => x.Key, x => x.Value)
-                                .ToDictionary(x => x.Key, g => g.First());
-            }
-
-            string customIniContent = String.Empty;
-
-            foreach (KeyValuePair<string, string> item in customWeb) {
-                customIniContent += item.Key + " = " + item.Value + "\n";
-            }
-
-            File.WriteAllText(iniSettingsFolder + "/custom_web.ini", customIniContent);
-            customFilesWithPath.Add("web", iniSettingsFolder + "/custom_web.ini");
-
-            // Combine config and configCLI to a custom.ini for CLI
-            var configCLI = new Dictionary<string, string>(PhpConfig.ConfigCli);
-            var customCLI = new Dictionary<string, string>(PhpConfig.Config);
-
-            foreach (KeyValuePair<string, string> item in customCLI) {
-                if (configCLI.ContainsKey(item.Key)) {
-                    customCLI[item.Key] = configCLI[item.Key];
-                    configCLI.Remove(item.Key);
-                }
-            }
-
-            if (configCLI.Count > 0) {
-                customCLI = customCLI.Union(configCLI)
-                                .ToLookup(x => x.Key, x => x.Value)
-                                .ToDictionary(x => x.Key, g => g.First());
-            }
-
-            customIniContent = String.Empty;
-
-            foreach (KeyValuePair<string, string> item in customCLI) {
-                customIniContent += item.Key + " = " + item.Value + "\n";
-            }
-
-            File.WriteAllText(iniSettingsFolder + "/custom_cli.ini", customIniContent);
-            customFilesWithPath.Add("cli", iniSettingsFolder + "/custom_cli.ini");
-
-            return customFilesWithPath;
         }
     }
 }
